+++ src/app/utils/ses.js
@@ -1,0 +1,8 @@
+import {SendEmailCommand, SESClient} from "@aws-sdk/client-ses";
+const sesClient = new SESClient();
+
+
+export function  sesSendEmailComand(params){
+    const sendEmailComand = new SendEmailCommand(params)
+    return sesClient.send(sendEmailComand);
+}
