+++ src/app/utils/sqs.js
@@ -1,0 +1,8 @@
+import {SendMessageCommand, SQSClient} from "@aws-sdk/client-sqs";
+
+const sqsClient = new SQSClient();
+
+
+export function sqsSendMessage(params){
+    return sqsClient.send(new SendMessageCommand(params));
+}
