--- src/ses_queue_handler/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "ses_queue_handler",
-  "version": "1.0.0",
-  "description": "",
-  "type": "module",
-  "main": "SesNotificationHandler.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Asif Hamza",
-  "license": "MIT-0",
-  "devDependencies": {
-    "@aws-sdk/client-dynamodb": "^3.341.0",
-    "@aws-sdk/lib-dynamodb": "^3.341.0"
-  }
-}
