+++ src/ses_queue_send_handler/handler.test.js
@@ -1,4 +1,4 @@
-import {getFieldFromManifiesto} from "./uitls.js";
+import {getFieldFromManifiesto} from "../app/utils/uitls.js";
 
 
 const campos = getFieldFromManifiesto({
