+++ template.yaml
@@ -15,6 +15,8 @@ Globals:
     MemorySize: 128
     Timeout: 10
 Resources:
+
+
   SESEventsTable:
     Type: 'AWS::DynamoDB::Table'
     Properties:
@@ -148,7 +150,7 @@ Resources:
   ProcessSESQueue:
     Type: 'AWS::Serverless::Function'
     Properties:
-      CodeUri: src/ses_queue_handler/
+      CodeUri: src/app/
       Handler: SesNotification.handler
       Environment:
         Variables:
@@ -170,7 +172,7 @@ Resources:
   SendNotificationFunction:
     Type: 'AWS::Serverless::Function'
     Properties:
-      CodeUri: src/ses_queue_send_handler/
+      CodeUri: src/app/
       Handler: SendNotification.handler
       Environment:
         Variables:
@@ -203,7 +205,7 @@ Resources:
   GetEventsFunction:
     Type: 'AWS::Serverless::Function'
     Properties:
-      CodeUri: src/get_events/
+      CodeUri: src/app/
       Handler: GetEvents.handler
       Environment:
         Variables:
